Issue #91 - Refactor search-issues.js to utilize centralized HTTP pagination logic
- Replaced inline pagination functions with `createHttpPaginators` utility to streamline the code and reduce duplication across different pipeline versions.
- This change enhances maintainability and consistency in handling paginated API responses for issue searches.
