+++ src/pipelines/sq-10.0/sonarcloud/api/issues/helpers/search-issues.js
@@ -1,7 +1,7 @@
 // -------- Search Issues --------
 
 import logger from '../../../../../../shared/utils/logger.js';
-import { fetchWithSlicing } from '../../../../../../shared/utils/search-slicer/index.js';
+import { fetchWithSlicing, createHttpPaginators } from '../../../../../../shared/utils/search-slicer/index.js';
 
 // SonarCloud only accepts the classic issue statuses for the `statuses` parameter.
 const ALL_STATUSES = 'OPEN,CONFIRMED,REOPENED,RESOLVED,CLOSED';
@@ -16,26 +16,6 @@ export async function searchIssues(client, organization, projectKey, filters = {
     ...filters,
   };
 
-  const probeTotalFn = async (endpoint, params) => {
-    const response = await client.get(endpoint, { params: { ...params, ps: 1, p: 1 } });
-    return response.data.paging?.total ?? 0;
-  };
-
-  const getPaginatedFn = async (endpoint, params, dataKey) => {
-    let allResults = [];
-    let page = 1;
-    const pageSize = 500;
-    // eslint-disable-next-line no-constant-condition
-    while (true) {
-      const response = await client.get(endpoint, { params: { ...params, ps: pageSize, p: page } });
-      const items = response.data[dataKey] || [];
-      allResults = allResults.concat(items);
-      const total = response.data.paging?.total || 0;
-      if (page * pageSize >= total || items.length < pageSize) break;
-      page++;
-    }
-    return allResults;
-  };
-
+  const { probeTotalFn, getPaginatedFn } = createHttpPaginators(client);
   return await fetchWithSlicing(probeTotalFn, getPaginatedFn, '/api/issues/search', baseParams, 'issues');
 }
