+++ src/pipelines/sq-9.9/sonarcloud/api/issues/helpers/search-issues.js
@@ -1,5 +1,5 @@
 import logger from '../../../../../../shared/utils/logger.js';
-import { fetchWithSlicing } from '../../../../../../shared/utils/search-slicer/index.js';
+import { fetchWithSlicing, createHttpPaginators } from '../../../../../../shared/utils/search-slicer/index.js';
 
 // -------- Search Issues with Pagination --------
 
@@ -15,26 +15,6 @@ export async function searchIssues(client, organization, projectKey, filters = {
     ...filters,
   };
 
-  const probeTotalFn = async (endpoint, params) => {
-    const response = await client.get(endpoint, { params: { ...params, ps: 1, p: 1 } });
-    return response.data.paging?.total ?? 0;
-  };
-
-  const getPaginatedFn = async (endpoint, params, dataKey) => {
-    let allResults = [];
-    let page = 1;
-    const pageSize = 500;
-    // eslint-disable-next-line no-constant-condition
-    while (true) {
-      const response = await client.get(endpoint, { params: { ...params, ps: pageSize, p: page } });
-      const items = response.data[dataKey] || [];
-      allResults = allResults.concat(items);
-      const total = response.data.paging?.total || 0;
-      if (page * pageSize >= total || items.length < pageSize) break;
-      page++;
-    }
-    return allResults;
-  };
-
+  const { probeTotalFn, getPaginatedFn } = createHttpPaginators(client);
   return await fetchWithSlicing(probeTotalFn, getPaginatedFn, '/api/issues/search', baseParams, 'issues');
 }
